+++ shapekey_splitter/operators/__init__.py
@@ -1,12 +1,13 @@
 if "bpy" in locals():
     import importlib
-    from . import op_split, op_mirror_weights, op_preview, op_masks
+    from . import op_split, op_mirror_weights, op_preview, op_masks, op_keys
+    importlib.reload(op_keys)
     importlib.reload(op_split)
     importlib.reload(op_mirror_weights)
     importlib.reload(op_preview)
     importlib.reload(op_masks)
 else:
-    from . import op_split, op_mirror_weights, op_preview, op_masks
+    from . import op_split, op_mirror_weights, op_preview, op_masks, op_keys
 
 import bpy
 
@@ -21,6 +22,9 @@ classes = (
     op_masks.SHAPEKEY_OT_mask_remove,
     op_masks.SHAPEKEY_OT_mask_rename,
     op_masks.SHAPEKEY_OT_select_vertices_for_paint,
+    op_keys.SHAPEKEY_OT_key_toggle,
+    op_keys.SHAPEKEY_OT_keys_select_all,
+    op_keys.SHAPEKEY_OT_keys_clear_selection,
 )